move qs decoder and parser to its own file
